moved some engraving draw services to internal dir

+++ src/engraving/draw/utils/drawcomp.cpp
@@ -216,7 +216,7 @@ static bool isEqual(const PainterPath& v1, const PainterPath& v2, double toleran
         return false;
     }
 
-    for (int i = 0; i < v1.elementCount(); ++i) {
+    for (size_t i = 0; i < v1.elementCount(); ++i) {
         PainterPath::Element e1 = v1.elementAt(i);
         PainterPath::Element e2 = v2.elementAt(i);
